Improve structure of tests.
